extract autocomplete
